Move Linux IOPriority
